+++ src/common/services/breathecode.js
@@ -140,10 +140,16 @@ const breathecode = {
       get: () => axios.get(`${url}/me`),
     };
   },
-  public: () => {
+
+  public: (query = {}) => {
     const url = `${host}/admissions/public`;
+
+    const qs = Object.keys(query)
+      .map((key) => (query[key] !== undefined ? `${key}=${query[key]}` : ''))
+      .join('&');
     return {
-      mentors: () => axios.get(`${url}/cohort/user?syllabus=coding-introduction&roles=TEACHER,ASSISTANT`),
+      mentors: () => axios.get(`${url}/cohort/user?${qs}`),
+      events: () => axios.get(`${host}/events/all?${qs}`),
     };
   },
 };
